test(task-migrate-bulk): 迁 import_export 的 68 个无帧用例到 gui_unit_test（清单上界 61 → 实测 68）
test_gui_import_export.cpp 的 89 个用例里，68 个是纯序列化往返：构造 GuiState、
走生产路径序列化、读回、对文档断言。它们不需要窗口、GL 上下文或 ImGuiTestContext，
而 gui_test 在 CI 里是 build-only（无显示服务器）——也就是说这 68 个从未在
Windows 或 Linux 上跑过，而序列化恰是平台差异（浮点格式化、路径分隔符、locale）
真正发作的地方。

留在 gui_test 的 21 个：GL / 文档重置 owner 簇（接缝是「文档切换 →
ResetFrontendState → 预览纹理 / poller staged」）+ 经 GuiFunc/YieldUntilTrue 推帧的用例。

三点与清单不一致，均为实测修正：

* 68 > issue 上界 61。静态清单按「匿名 ctx」+「IM_UNUSED(ctx) 标记」两种形态计数，
  漏掉第三种：具名 ctx 参数、通篇不用、也不带 IM_UNUSED（6 个）。
* GL 簇不是连续 14 个。该行段内 12 个真依赖 GL/ctx，另 2 个是纯 CPU：
  reset_primitives_are_owner_single_source 只读 app.cpp 源码文本做计数断言，
  dorevert_invalidates_effects_baselines_via_owner 只做结构体赋值 + DoRevert。
  两者不共享 fixture，随逻辑侧迁走；前者本身是门禁，留在 gui_test 等于它永不在 CI 跑。
* IM_CHECK* 映射为 ASSERT_* 而非 EXPECT_*。imgui_te_context.h:568 的 IM_CHECK 与
  :592 的 IM_CHECK_OP（_RETURN=true）在失败时都 `return`，即 fail-fast 语义；
  这两个文件一次都没用过带 EXPECT 语义的 *_NO_RET 变体。用 EXPECT_* 会让失败的守卫
  继续往下跑进依赖它的代码（contains 之后取下标、判空之后解引用）。

SceneJson / CommitSceneJson / CoreJson / PrismFacePlaneOffsets / CountDistinct 两侧都要用，
提升到 test/support/scene_json_helpers.hpp（跟随 support/user_defaults_test_env.hpp 先例，
两个 target 的 include 路径都已含 test/ 根，无需改 CMake include-dir）。

随迁的注释里对 scratchpad 工作制品的引用就地改写为机制本身（check_new_refs.py 是
diff-scoped 的，迁移会把陈年散文重新算作新增行；那些引用确实悬空）。

核验：断言按宏形态逐一守恒（IM_CHECK 262 = 留 76 + 迁 186，IM_CHECK_EQ 319 = 104 + 215，
_STR_EQ 36 = 0 + 36，_LT 16 = 0 + 16，_NE 1 = 0 + 1）；gui_unit_test 156 → 224 全绿。
